добавил liquibase

+++ src/main/java/com/example/newjiraback/controller/TodoController.java
@@ -27,12 +27,12 @@ public class TodoController {
     }
 
     @GetMapping("/get/{id}")
-    public Todo get(@PathVariable int id) {
+    public Todo get(@PathVariable long id) {
         return todoService.get(id);
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable long id) {
         todoService.delete(id);
     }
